requirements.txt
openai-whisper>=20231117
openai>=1.0.0
edge-tts>=6.1.0
demucs>=4.0.0
pydub>=0.25.1
yt-dlp>=2024.1.1
